fix(trusty-mpm): WI-3 final — forward full batch around AuthPrompt, tighten git-op bracket SHA, widen restart_count to u32 (refs #1594)
- Remove break from TrySendError::Closed branch inside the event-forwarding
  for loop: the Closed path previously exited the loop early, silently
  dropping broadcast sends for remaining batch events; now critical_tx is
  cleared and the loop continues so every event in the batch reaches broadcast
  regardless of critical-observer state or AuthPrompt position
- Tighten git-op bracket SHA pattern from `[0-9a-f]+` (1+ chars) to
  `[0-9a-f]{7,40}` so non-git bracket tokens like `[session a1b2c3]` (6 chars)
  or `[error 404abc]` (6 chars) no longer classify as GitOp; real 7–40 char
  SHAs still match
- Widen CtlSessionSummary.restart_count from u8 to u32 at the API boundary
  (cast from SessionMetadata.restart_count as u32) so crash-looping sessions
  report real counts beyond 255
- Add actor_auth_prompt_batch_trailing_event_forwarded: asserts a trailing
  Output event after an AuthPrompt in the same batch is still broadcast
- Add activity_parser_git_op_bracket_sha_length: table-driven test with
  false-positive cases (6-char tokens) and true-positive cases (7+char SHAs)

+++ crates/trusty-mpm/src/daemon/api/control_routes.rs
@@ -189,7 +189,10 @@ pub struct CtlSessionSummary {
     /// State label.
     pub state: String,
     /// Number of times the backend has been auto-restarted (§3.1).
-    pub restart_count: u8,
+    /// Widened to u32 at the API boundary so crash-looping sessions never
+    /// saturate at 255 (the source field SessionMetadata::restart_count is
+    /// still u8 per spec; we cast with `as u32` here).
+    pub restart_count: u32,
     /// `true` if the session has restarted ≥1 time (no history replay in alpha-1).
     pub context_lost: bool,
     /// Seconds since the session started.
@@ -427,7 +430,7 @@ pub async fn ctl_list_sessions(
                 project_id: meta.project_id.clone(),
                 backend: backend_label.to_owned(),
                 state: meta.state.label().to_owned(),
-                restart_count: meta.restart_count,
+                restart_count: meta.restart_count as u32,
                 context_lost: meta.context_lost,
                 uptime_secs: meta.uptime_secs(),
                 last_activity_ts: meta.last_activity_at.to_rfc3339(),
